requirements.txt
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
numpy>=1.24.0
pyyaml>=6.0
click>=8.1.3
tqdm>=4.65.0
colorama>=0.4.6
requests>=2.28.2
cachetools>=5.3.0
diskcache>=5.6.1
pytest>=7.3.1
streamlit>=1.22.0  # Optional for UI
python-dotenv>=1.0.0
spacy>=3.5.0
thefuzz>=0.19.0
plotly>=5.13.0
pandas>=1.5.3
transformers>=4.28.1
#en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
